% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mclust_SEP.R
\name{Mclust_SEP_cpp}
\alias{Mclust_SEP_cpp}
\title{The model-based clustering for longitudinal data}
\usage{
Mclust_SEP_cpp(Y, p, q, Ks, type_cov, tol = 0.001, maxit = 500,
  save_fit = TRUE)
}
\arguments{
\item{Y}{A r-by-(\code{p}*\code{q}) matrix where r is the sample size, and ordering of columns should be carefully set (see `Details').}

\item{p, q}{An integer value for the number of multi-variables and the number of time points, respectively.}

\item{Ks}{A sequence of positive integers indicating the number of mixture components, each of which will be used in \code{K} of \code{\link{Mclust_SEP_each_cpp}}.}

\item{type_cov}{A sequence of character strings indicating covariance structures, each of which will be used in \code{\link{Mclust_SEP_each_cpp}}. Default is \code{'all'}, which runs all available models. See `Details'.}

\item{tol}{Tolerance constant for convergence. Default is \code{1e-3}.}

\item{maxit}{Maximum number of iterations. Default is \code{500}.}

\item{save_fit}{A logical value indicating whether to save all fitted mixture models. If \code{FALSE}, the best model is only available by \code{best_model}. Default is \code{TRUE}.}
}
\value{
A list with components:
\item{best_model}{A list of the mixture models with the largest BIC. If there is a tie, they are all returned.}
\item{bic_table}{Table filled with BIC values. Type of covariance models are given by rows and values in \code{Ks} by columns.}
\item{res_Mclust_SEP}{If \code{save_fit} is TRUE, all fitted models are stored. It is a nested list with the first layer corresponding to covariance models specified in \code{type_cov} and the second to values of \code{Ks}.}
}
\description{
This is a wrapper function of \code{\link{Mclust_SEP_each_cpp}}. All arguments except \code{save_fit} will be passed to \code{\link{Mclust_SEP_each_cpp}}.
}
\details{
The first \code{q} components from each row of \code{Y} denote \code{q} variables at time point 1, the second \code{q} are those at time point 2, and so on until time point \code{p}. Under separability, the covariance matrix of row vectors of \code{Y} is represented by \eqn{U_{p\times p} \otimes V_{q \times q}} for some covariance factors \eqn{U_{p\times p}, V_{q \times q}}.\cr\cr
\code{type_cov} should be in ``XXX-YYY'' format. ``XXX'' is for the multivariable covariance \eqn{U_{p\times p}}, and ``YYY'' for the temporal covariance \eqn{V_{q \times q}}. They will be passed respectively to \code{type_vari} and \code{type_time} in \code{\link{Mclust_SEP_each_cpp}}. Available options are as follows;
\itemize{
  \item Heteroscadatsic covariance structure : \code{VVV-VUN} (unstructured), \code{VVV-VAR} (AR), \code{VVV-VCS} (CS).
  \item Homoscadatsic covariance structure : \code{EEE-EUN} (unstructured), \code{EEE-EAR} (AR), \code{EEE-ECS} (CS).
}
For initialization of cluster membership, see `Details' in \code{\link{Mclust_SEP_each_cpp}}.
}
\examples{
# Gaussian mixture model with two components
K <- 2
p <- 2
q <- 3
U <- lapply(1:K, function(noarg) getCovariance(p, 0.3, "AR")) # getCovariance in Rcpp
V <- lapply(1:K, function(noarg) getCovariance(q, 0.2, "CS"))
Sigma <- Map(kronecker, U, V) # separable covariance matrix
mu <- list(rep(0, p * q), 5 / sqrt(p*q) * rep(1, p * q)) # distinct mean vectors
Y <- vector(mode = "list", length = K)
for(i in 1:K){
  Y[[i]] <- mvtnorm::rmvnorm(n = 20, mean = mu[[i]], sigma = Sigma[[i]])
}
fit <- Mclust_SEP_cpp(Y = Reduce(rbind, Y), p = p, q = q, Ks = 2, type_cov = "EEE-ECS")
}
\seealso{
\code{\link{Mclust_SEP_each_cpp}}
}
