\name{MclustSepCov-package}
\alias{MclustSepCov-package}
\alias{MclustSepCov}
\docType{package}
\title{
\packageTitle{MclustSepCov}
}
\description{
\packageDescription{MclustSepCov}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MclustSepCov}
\packageIndices{MclustSepCov}
}
\author{
\packageAuthor{MclustSepCov}

Maintainer: \packageMaintainer{MclustSepCov}
}
\keyword{ package }
